ivp_collision: Ensure enum sum is enum (559c566)
- Add explicit int casts when summing IVP_BOOL enum values

+++ ivp_collision/ivp_mindist_mcases.cxx
@@ -576,7 +576,8 @@ IVP_MRC_TYPE IVP_Mindist_Minimize_Solver::p_minimize_Leave_KK(const IVP_Compact_
         m_cache_L->tmp.synapse->update_synapse(L, IVP_ST_EDGE);
 
         // End KL	// sk and sl are already set by master
-        if (reverse_side_check[0] + reverse_side_check[1] == IVP_TRUE + IVP_TRUE)
+        // dimhotepus: Ensure enum sum is enum.
+        if ((int)reverse_side_check[0] + (int)reverse_side_check[1] == (int)IVP_TRUE + (int)IVP_TRUE)
         {
             m_cache_K->tmp.synapse->update_synapse(K, IVP_ST_BACKSIDE);
             m_cache_L->tmp.synapse->update_synapse(L, IVP_ST_TRIANGLE);
@@ -584,7 +585,8 @@ IVP_MRC_TYPE IVP_Mindist_Minimize_Solver::p_minimize_Leave_KK(const IVP_Compact_
             this->pos_opposite_BacksideOs.set_interpolate(&L_Kos[0], &L_Kos[1], sl);
             return IVP_MRC_BACKSIDE;
         }
-        if (reverse_side_check[2] + reverse_side_check[3] == IVP_TRUE + IVP_TRUE)
+        // dimhotepus: Ensure enum sum is enum.
+        if ((int)reverse_side_check[2] + (int)reverse_side_check[3] == (int)IVP_TRUE + (int)IVP_TRUE)
         {
             m_cache_K->tmp.synapse->update_synapse(K, IVP_ST_TRIANGLE);
             IVP_DOUBLE sk = kkr->checks_K[0] / (kkr->checks_K[0] + kkr->checks_K[1]);
